feat: convert screens, navigation, and entry points to TypeScript
Type all screens with NativeStackScreenProps<RootStackParamList, ScreenName>;
type AppNavigator with createNativeStackNavigator<RootStackParamList>(); replace
animationEnabled (removed in newer @react-navigation/native-stack) with
animation:'none'; rename App.jsx → App.tsx and index.js → index.ts.
