+++ App.tsx
@@ -1,9 +1,8 @@
 import { StatusBar } from 'expo-status-bar';
-import "./global.css";
+import './global.css';
 import { AppNavigator } from './src/navigation/AppNavigator';
 
 export default function App() {
-
   return (
     <>
       <StatusBar style="dark" />
