+++ index.ts
@@ -1,5 +1,4 @@
 import { registerRootComponent } from 'expo';
 import App from './App';
 
-// Register the main component for Expo
 registerRootComponent(App);
