+++ src/screens/ProfileScreen.tsx
@@ -1,86 +1,77 @@
 import React, { useEffect, useState } from 'react';
 import { Alert, View } from 'react-native';
 import { useNavigation } from '@react-navigation/native';
+import type { NavigationProp } from '@react-navigation/native';
 import Screen from '../components/layout/Screen';
 import ProfileForm from '../components/forms/ProfileForm';
 import NotificationForm from '../components/forms/NotificationForm';
 import { Button } from '../components/ui/Button';
 import useProfile from '../hooks/useProfile';
 import useAuth from '../hooks/useAuth';
+import type { RootStackParamList, UserProfile, NotificationPreferences } from '../types';
 
 const ProfileScreen = () => {
-  const navigation = useNavigation();
-  const { 
-    profile, 
-    preferences, 
-    loading, 
-    loadProfile, 
-    saveProfile, 
-    updateProfile, 
+  const navigation = useNavigation<NavigationProp<RootStackParamList>>();
+  const {
+    profile,
+    preferences,
+    loading,
+    loadProfile,
+    saveProfile,
+    updateProfile,
     updatePreferences,
     pickImage,
-    removeAvatar
+    removeAvatar,
   } = useProfile();
-  
+
   const { logout } = useAuth();
   const [hasChanges, setHasChanges] = useState(false);
-  
-  // Load profile data when component mounts
+
   useEffect(() => {
     loadProfile();
   }, []);
-  
-  // Track whether form has changed
+
   useEffect(() => {
     setHasChanges(true);
   }, [profile, preferences]);
-  
-  const handleSaveProfile = async (formData) => {
+
+  const handleSaveProfile = async (formData: Partial<UserProfile>) => {
     try {
-      // Update profile with form data
       updateProfile(formData);
-      
-      // Save to Firebase
       await saveProfile();
-      
       setHasChanges(false);
       Alert.alert('Success', 'Your changes have been saved');
-    } catch (error) {
+    } catch {
       Alert.alert('Error', 'Failed to save profile changes');
     }
   };
-  
-  const handleNotificationChange = (newPreferences) => {
+
+  const handleNotificationChange = (newPreferences: NotificationPreferences) => {
     updatePreferences(newPreferences);
   };
-  
+
   const handleDiscardChanges = async () => {
     await loadProfile();
     setHasChanges(false);
   };
-  
+
   const handleLogout = async () => {
-    Alert.alert(
-      'Confirm Logout',
-      'Are you sure you want to log out?',
-      [
-        { text: 'Cancel', style: 'cancel' },
-        {
-          text: 'Logout',
-          style: 'destructive',
-          onPress: async () => {
-            try {
-              await logout(navigation);
-              // Navigation is handled in the logout function
-            } catch (error) {
-              // Handle error silently
-            }
-          },
+    Alert.alert('Confirm Logout', 'Are you sure you want to log out?', [
+      { text: 'Cancel', style: 'cancel' },
+      {
+        text: 'Logout',
+        style: 'destructive',
+        onPress: async () => {
+          try {
+            await logout(navigation);
+          } catch {
+            // handled silently
+          }
         },
-      ]
-    );
+      },
+    ]);
   };
-  
+
   return (
     <Screen
       title="Profile"
@@ -89,7 +80,7 @@ const ProfileScreen = () => {
         firstName: profile.firstName,
         lastName: profile.lastName,
         avatar: profile.avatar,
-        showAvatar: true
+        showAvatar: true,
       }}
     >
       <ProfileForm
@@ -99,13 +90,7 @@ const ProfileScreen = () => {
         onRemoveImage={removeAvatar}
         loading={loading}
       />
-      
-      <NotificationForm
-        preferences={preferences}
-        onChange={handleNotificationChange}
-      />
-      
-      {/* Logout Button */}
+      <NotificationForm preferences={preferences} onChange={handleNotificationChange} />
       <Button
         title="Logout"
         onPress={handleLogout}
@@ -113,8 +98,6 @@ const ProfileScreen = () => {
         className="mt-6 mb-4"
         fullWidth
       />
-      
-      {/* Save/Discard Buttons */}
       {hasChanges && (
         <View className="flex-row justify-between mb-8">
           <Button
